+++ deepinv/models/__init__.py
@@ -43,8 +43,8 @@ from .wrapper import (
     DiffusersDenoiserWrapper,
     ComplexDenoiserWrapper,
     MinusOneOneDenoiserWrapper,
-    AnscombeDenoiserWrapper,
 )
+from .anscombe import AnscombeDenoiser
 from .poisson2sparse import ConvLista, Poisson2Sparse
 from .kernel_network import KernelIdentificationNetwork
 from .bilateral import BilateralFilter
