+++ deepinv/models/wrapper.py
@@ -4,7 +4,6 @@ from torch import Tensor, nn
 from deepinv.models import Denoiser
 from typing import Callable
 import numpy as np
-from .anscombe import generalized_anscombe_transform, inverse_generalized_anscombe_transform
 
 
 class ScoreModelWrapper(Denoiser):
@@ -673,90 +672,3 @@ class MinusOneOneDenoiserWrapper(nn.Module):
         if not kwargs.get("input_in_minus_one_one", True):
             denoised = 2 * (denoised - self.xmin) / (self.xmax - self.xmin) - 1
         return denoised
-
-
-class AnscombeDenoiserWrapper(Denoiser):
-    r"""
-    Wraps a Gaussian denoiser into a Poisson-Gaussian denoiser using the
-    :func:`Generalized Anscombe Transform (GAT) <deepinv.models.generalized_anscombe_transform>`
-    and its :func:`inverse <deepinv.models.inverse_generalized_anscombe_transform>`.
-
-    Given a noisy observation :math:`y` corrupted by
-    :class:`Poisson-Gaussian noise <deepinv.physics.PoissonGaussianNoise>` with gain :math:`\gamma`
-    and Gaussian standard deviation :math:`\sigma`, the wrapper:
-
-    1. Applies the GAT to stabilize the variance to approximately :math:`\gamma^2`:
-
-    .. math::
-
-        z = 2 \sqrt{\gamma y + \frac{3}{8}\gamma^2 + \sigma^2}
-
-    2. Applies the wrapped Gaussian denoiser :math:`\denoisername` at noise level :math:`\gamma`
-       (which matches the standard deviation of the GAT output):
-
-    .. math::
-
-        \hat{z} = \denoisername(z,\; \gamma)
-
-    3. Applies the inverse GAT to return to the original domain.
-       Setting :math:`u = \hat{z}/\gamma`:
-
-    .. math::
-
-        \hat{y} = \gamma\left(\frac{1}{4}u^2 + \frac{1}{4}\sqrt{\frac{3}{2}}\,u^{-1} - \frac{11}{8}u^{-2} + \frac{5}{8}\sqrt{\frac{3}{2}}\,u^{-3} - \frac{1}{8} + \frac{\sigma^2}{\gamma^2}\right), \qquad u = \frac{\hat{z}}{\gamma}
-
-    .. note::
-
-        When ``gain = 0`` the noise is purely Gaussian and the GAT/IGAT are bypassed:
-        the wrapped denoiser is called directly as :math:`\denoisername(y, \sigma)`.
-
-    |sep|
-
-    :Examples:
-
-        >>> import deepinv as dinv
-        >>> import torch
-        >>> from deepinv.models import AnscombeDenoiserWrapper, DRUNet
-        >>> denoiser = DRUNet(pretrained=None)
-        >>> anscombe_denoiser = AnscombeDenoiserWrapper(denoiser, sigma_denoiser=25/255)
-        >>> y = torch.rand(1, 3, 32, 32)
-        >>> with torch.no_grad():
-        ...     y_denoised = anscombe_denoiser(y, gain=0.1, sigma=0.05)
-
-    :param deepinv.models.Denoiser denoiser: Gaussian denoiser :math:`\denoisername` to wrap.
-    """
-
-    def __init__(self, denoiser: Denoiser,  *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.denoiser = denoiser
-
-    def forward(
-        self,
-        y: torch.Tensor,
-        gain: float | torch.Tensor,
-        sigma: float | torch.Tensor,
-        *args,
-        **kwargs,
-    ) -> torch.Tensor:
-        r"""
-        Applies the Poisson-Gaussian denoiser via the Anscombe transform.
-
-        :param torch.Tensor y: Noisy measurements corrupted by Poisson-Gaussian noise.
-        :param float | torch.Tensor gain: Gain of the Poisson distribution :math:`\gamma`.
-        :param float | torch.Tensor sigma: Standard deviation of the Gaussian noise :math:`\sigma`.
-        :param args: Additional positional arguments passed to the wrapped denoiser.
-        :param kwargs: Additional keyword arguments passed to the wrapped denoiser.
-        :return torch.Tensor: Denoised measurements in the original domain.
-        """
-        # Bypass GAT/IGAT for purely Gaussian noise (gain = 0)
-        if (isinstance(gain, torch.Tensor) and gain == 0) or gain == 0:
-            return self.denoiser(y, sigma, *args, **kwargs)
-
-        # 1. Apply GAT: stabilize variance to ~γ²
-        z = generalized_anscombe_transform(y, gain, sigma)
-
-        # 2. Denoise in the Anscombe domain at noise level σ = γ (matching the GAT output std)
-        z_denoised = self.denoiser(z, sigma=gain, *args, **kwargs)
-
-        # 3. Apply inverse GAT to return to original domain
-        return inverse_generalized_anscombe_transform(z_denoised, gain, sigma)
